% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_dump.R
\name{download_data_dump}
\alias{download_data_dump}
\title{Read the ROR data dump}
\usage{
download_data_dump(filepath)
}
\arguments{
\item{filepath}{Filepath to save .zip file to}
}
\value{
A list of dataframes holding the data dump.
}
\description{
In addition to their API, ROR also provide a data dump via Zenodo (updated ~quarterly). It is stored in .zip format. This function saves to a temp directory and then turns the .json file within into a dataframe. NOTE THE FILE IS LARGE: >1GB.
}
