% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id_map.R
\name{get_id_map}
\alias{get_id_map}
\title{Get a map of different organisation IDs.}
\usage{
get_id_map(id)
}
\arguments{
\item{id}{The unique, non-ROR ID for the organisation concerned.}
}
\value{
A list of any available IDs (or a message confirming a match could not be found).
}
\description{
ROR holds multiple types of unique ID for organisations. This function let's you enter one such ID and return all the other IDs in the database for that organisation. NOTE THIS FUNCTION IS ONLY RECOMMENDED IF YOU HAVE A RELATIVELY SMALL NUMBER OF IDs TO MAP. Otherwise use \code{read_data_dump()} to access a snapshot of the full dataset.
}
