% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_org_list.R
\name{get_org_list}
\alias{get_org_list}
\title{Query organisation records to return a list; you can add query terms to refine results.}
\usage{
get_org_list(
  .search_terms,
  .affiliation,
  .org_type,
  .country_code,
  .country_name,
  .page_numbers,
  .include_non_active = "false"
)
}
\arguments{
\item{.search_terms}{Query term to narrow results if desired. Separate multiple terms with "|" to search on an OR basis. Will search all indexed fields.}

\item{.affiliation}{Specific search term for an affiliation. The API determines closeness of match - only those deemed close enough ('chosen') returned. Will only search name, label and alias fields. You cannot combine affiliation searches with filters.}

\item{.org_type}{Only return specified type of organisation. Any of "Education", "Healthcare", "Company", "Archive", "Nonprofit", "Government", "Facility" or "Other".}

\item{.country_code}{Filter to specific countries as per the ISO 3166 alpha-2 list. Can combine with \code{.org_type} filter.}

\item{.country_name}{Alternatively, filter by country code instead of name. Can combine with \code{.org_type} filter.}

\item{.page_numbers}{Return a specific page number (or set of page numbers). Enter as a numeric vector. All results will be returned if you do not specify. If you specify page numbers that are not found, all results will be returned.}

\item{.include_non_active}{Option to include 'inactive' or 'withdrawn' entities. Must be 'true' or 'false'.}
}
\value{
A dataframe holding results as per specifications (list of all orgs returned if no parameters are set).
}
\description{
By defaults returns all results from the API. Also let's you specify a specific page number, filter your results by specific categories, or add search terms to your query to refine the results. It is HIGHLY RECOMMENDED you filter your results in some way to minimise running times.
}
