% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results_by_page.R
\name{get_results_by_page}
\alias{get_results_by_page}
\title{Helper function - returns org list details one page at a time..}
\usage{
get_results_by_page(
  ..page_num = 1,
  ..search_terms,
  ..affiliation,
  ..org_type,
  ..country_code,
  ..country_name,
  ..include_non_active = "false"
)
}
\arguments{
\item{..page_num}{Return a specific page number. Enter as an integer - if not set, defaults to 1.}

\item{..search_terms}{Query term to narrow results if desired. Separate multiple terms with "|" to search on an OR basis. Will search all indexed fields.}

\item{..affiliation}{Specific search term for an affiliation. The API determines closeness of match - only those deemed close enough ('chosen') returned. Will only search name, label and alias fields. You cannot combine affiliation searches with filters.}

\item{..org_type}{Only return specified type of organisation. Any of "Education", "Healthcare", "Company", "Archive", "Nonprofit", "Government", "Facility" or "Other".}

\item{..country_code}{Filter to specific countries as per the ISO 3166 alpha-2 list. Can combine with \code{..org_type} filter.}

\item{..country_name}{Alternatively, filter by country code instead of name. Can combine with \code{..org_type} filter.}

\item{..include_non_active}{Option to include 'inactive' or 'withdrawn' entities. Must be 'true' or 'false'.}
}
\value{
A dataframe holding the first page of results from the API.
}
\description{
Created as a helper function to return org list results one page at a time. Can be used in its own right, however.
}
