% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ror-package.R
\docType{package}
\name{ror-package}
\alias{ror}
\alias{ror-package}
\title{ror: Access The Registry of Research (ROR) API Via R}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

An R wrapper for interacting with the Registry of Research API Allows you to return individual organisation records, a list of all organisations, or search by specific terms.
}
\keyword{internal}
